int16_t inventory

+++ inc/shared/game.h
@@ -208,7 +208,7 @@ typedef struct {
     void (*ClientPrint)(edict_t *ent, print_level_t printlevel, const char *msg);
     void (*ClientLayout)(edict_t *ent, const char *str, bool reliable);
     void (*ClientStuffText)(edict_t *ent, const char *str);
-    void (*ClientInventory)(edict_t *ent, int *inventory, int count);
+    void (*ClientInventory)(edict_t *ent, int16_t *inventory, int count);
 
     int (*DirToByte)(const vec3_t dir);
     void (*ByteToDir)(int v, vec3_t dir);
